Thread-safety test is for the underlying storage
